+++ cmd/builder/internal/command.go
@@ -24,9 +24,21 @@ import (
 	"github.com/knadh/koanf/providers/file"
 	"github.com/knadh/koanf/providers/posflag"
 	"github.com/spf13/cobra"
+	flag "github.com/spf13/pflag"
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/cmd/builder/internal/builder"
+)
+
+const (
+	skipCompilationFlag            = "skip-compilation"
+	distributionNameFlag           = "name"
+	distributionDescriptionFlag    = "description"
+	distributionVersionFlag        = "version"
+	distributionOtelColVersionFlag = "otelcol-version"
+	distributionOutputPathFlag     = "output-path"
+	distributionGoFlag             = "go"
+	distributionModuleFlag         = "module"
 )
 
 var (
@@ -38,73 +50,117 @@ var (
 // Command is the main entrypoint for this application
 func Command() (*cobra.Command, error) {
 	cmd := &cobra.Command{
-		Use:  "builder",
-		Long: fmt.Sprintf("OpenTelemetry Collector distribution builder (%s)", version),
+		SilenceUsage:  true, // Don't print usage on Run error.
+		SilenceErrors: true, // Don't print errors; main does it.
+		Use:           "ocb",
+		Long: fmt.Sprintf("OpenTelemetry Collector Builder (%s)", version) + `
+
+ocb generates a custom OpenTelemetry Collector binary using the
+build configuration given by the "--config" argument.
+`,
+		Args: cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := initConfig(); err != nil {
+			if err := initConfig(cmd.Flags()); err != nil {
 				return err
 			}
 			if err := cfg.Validate(); err != nil {
-				cfg.Logger.Error("invalid configuration", zap.Error(err))
-				return err
+				return fmt.Errorf("invalid configuration: %w", err)
 			}
 
 			if err := cfg.ParseModules(); err != nil {
-				cfg.Logger.Error("invalid module configuration", zap.Error(err))
-				return err
+				return fmt.Errorf("invalid module configuration: %w", err)
 			}
 
 			return builder.GenerateAndCompile(cfg)
 		},
 	}
 
-	// the external config file
-	cmd.Flags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.otelcol-builder.yaml)")
+	cmd.Flags().StringVar(&cfgFile, "config", "", "build configuration file")
+
+	// A build configuration file is always required, and there's no
+	// default. We can relax this in future by embedding the default
+	// config that is used to build otelcorecol.
+	if err := cmd.MarkFlagRequired("config"); err != nil {
+		panic(err) // Only fails if the usage message is empty, which is a programmer error.
+	}
 
 	// the distribution parameters, which we accept as CLI flags as well
-	cmd.Flags().BoolVar(&cfg.SkipCompilation, "skip-compilation", false, "Whether builder should only generate go code with no compile of the collector (default false)")
-	cmd.Flags().StringVar(&cfg.Distribution.Name, "name", "otelcol-custom", "The executable name for the OpenTelemetry Collector distribution")
-	cmd.Flags().StringVar(&cfg.Distribution.Description, "description", "Custom OpenTelemetry Collector distribution", "A descriptive name for the OpenTelemetry Collector distribution")
-	cmd.Flags().StringVar(&cfg.Distribution.Version, "version", "1.0.0", "The version for the OpenTelemetry Collector distribution")
-	cmd.Flags().StringVar(&cfg.Distribution.OtelColVersion, "otelcol-version", cfg.Distribution.OtelColVersion, "Which version of OpenTelemetry Collector to use as base")
-	cmd.Flags().StringVar(&cfg.Distribution.OutputPath, "output-path", cfg.Distribution.OutputPath, "Where to write the resulting files")
-	cmd.Flags().StringVar(&cfg.Distribution.Go, "go", "", "The Go binary to use during the compilation phase. Default: go from the PATH")
-	cmd.Flags().StringVar(&cfg.Distribution.Module, "module", "go.opentelemetry.io/collector/cmd/builder", "The Go module for the new distribution")
+	cmd.Flags().BoolVar(&cfg.SkipCompilation, skipCompilationFlag, false, "Whether builder should only generate go code with no compile of the collector (default false)")
+	cmd.Flags().StringVar(&cfg.Distribution.Name, distributionNameFlag, "otelcol-custom", "The executable name for the OpenTelemetry Collector distribution")
+	cmd.Flags().StringVar(&cfg.Distribution.Description, distributionDescriptionFlag, "Custom OpenTelemetry Collector distribution", "A descriptive name for the OpenTelemetry Collector distribution")
+	cmd.Flags().StringVar(&cfg.Distribution.Version, distributionVersionFlag, "1.0.0", "The version for the OpenTelemetry Collector distribution")
+	cmd.Flags().StringVar(&cfg.Distribution.OtelColVersion, distributionOtelColVersionFlag, cfg.Distribution.OtelColVersion, "Which version of OpenTelemetry Collector to use as base")
+	cmd.Flags().StringVar(&cfg.Distribution.OutputPath, distributionOutputPathFlag, cfg.Distribution.OutputPath, "Where to write the resulting files")
+	cmd.Flags().StringVar(&cfg.Distribution.Go, distributionGoFlag, "", "The Go binary to use during the compilation phase. Default: go from the PATH")
+	cmd.Flags().StringVar(&cfg.Distribution.Module, distributionModuleFlag, "go.opentelemetry.io/collector/cmd/builder", "The Go module for the new distribution")
 
 	// version of this binary
 	cmd.AddCommand(versionCommand())
 
 	if err := k.Load(posflag.Provider(cmd.Flags(), ".", k), nil); err != nil {
-		cfg.Logger.Error("failed to load command line arguments", zap.Error(err))
+		return nil, fmt.Errorf("failed to load command line arguments: %w", err)
 	}
 
 	return cmd, nil
 }
 
-func initConfig() error {
-	cfg.Logger.Info("OpenTelemetry Collector distribution builder", zap.String("version", version), zap.String("date", date))
-	// use the default path if there is no config file being specified
-	if cfgFile == "" {
-		cfgFile = "$HOME/.otelcol-builder.yaml"
-	}
+func initConfig(flags *flag.FlagSet) error {
+	cfg.Logger.Info("OpenTelemetry Collector Builder",
+		zap.String("version", version), zap.String("date", date))
 
 	// load the config file
 	if err := k.Load(file.Provider(cfgFile), yaml.Parser()); err != nil {
-		cfg.Logger.Error("failed to load config file", zap.String("config-file", cfgFile), zap.Error(err))
+		return fmt.Errorf("failed to load configuration file: %w", err)
 	}
 
 	// handle env variables
 	if err := k.Load(env.Provider("", ".", func(s string) string {
 		return strings.ReplaceAll(s, ".", "_")
 	}), nil); err != nil {
-		cfg.Logger.Error("failed to load env var", zap.Error(err))
+		return fmt.Errorf("failed to load environment variables: %w", err)
 	}
 
-	if err := k.UnmarshalWithConf("", &cfg, koanf.UnmarshalConf{Tag: "mapstructure"}); err != nil {
-		cfg.Logger.Error("failed to unmarshal config", zap.Error(err))
-		return err
+	cfgFromFile := builder.Config{}
+	if err := k.UnmarshalWithConf("", &cfgFromFile, koanf.UnmarshalConf{Tag: "mapstructure"}); err != nil {
+		return fmt.Errorf("failed to unmarshal configuration: %w", err)
 	}
+
+	applyCfgFromFile(flags, cfgFromFile)
 
 	cfg.Logger.Info("Using config file", zap.String("path", cfgFile))
 	return nil
+}
+
+func applyCfgFromFile(flags *flag.FlagSet, cfgFromFile builder.Config) {
+	cfg.Exporters = cfgFromFile.Exporters
+	cfg.Extensions = cfgFromFile.Extensions
+	cfg.Receivers = cfgFromFile.Receivers
+	cfg.Processors = cfgFromFile.Processors
+	cfg.Replaces = cfgFromFile.Replaces
+	cfg.Excludes = cfgFromFile.Excludes
+
+	if !flags.Changed(skipCompilationFlag) && cfgFromFile.SkipCompilation {
+		cfg.SkipCompilation = cfgFromFile.SkipCompilation
+	}
+	if !flags.Changed(distributionNameFlag) && cfgFromFile.Distribution.Name != "" {
+		cfg.Distribution.Name = cfgFromFile.Distribution.Name
+	}
+	if !flags.Changed(distributionDescriptionFlag) && cfgFromFile.Distribution.Description != "" {
+		cfg.Distribution.Description = cfgFromFile.Distribution.Description
+	}
+	if !flags.Changed(distributionVersionFlag) && cfgFromFile.Distribution.Version != "" {
+		cfg.Distribution.Version = cfgFromFile.Distribution.Version
+	}
+	if !flags.Changed(distributionOtelColVersionFlag) && cfgFromFile.Distribution.OtelColVersion != "" {
+		cfg.Distribution.OtelColVersion = cfgFromFile.Distribution.OtelColVersion
+	}
+	if !flags.Changed(distributionOutputPathFlag) && cfgFromFile.Distribution.OutputPath != "" {
+		cfg.Distribution.OutputPath = cfgFromFile.Distribution.OutputPath
+	}
+	if !flags.Changed(distributionGoFlag) && cfgFromFile.Distribution.Go != "" {
+		cfg.Distribution.Go = cfgFromFile.Distribution.Go
+	}
+	if !flags.Changed(distributionModuleFlag) && cfgFromFile.Distribution.Module != "" {
+		cfg.Distribution.Module = cfgFromFile.Distribution.Module
+	}
 }