+++ service/extensions/extensions_test.go
@@ -93,13 +93,12 @@ func TestBuildExtensions(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := Build(context.Background(), Settings{
-				Telemetry:         componenttest.NewNopTelemetrySettings(),
-				BuildInfo:         component.NewDefaultBuildInfo(),
-				Configs:           tt.extensionsConfigs,
-				Factories:         tt.factories.Extensions,
-				ServiceExtensions: tt.serviceExtensions,
-			})
+			_, err := New(context.Background(), Settings{
+				Telemetry: componenttest.NewNopTelemetrySettings(),
+				BuildInfo: component.NewDefaultBuildInfo(),
+				Configs:   tt.extensionsConfigs,
+				Factories: tt.factories.Extensions,
+			}, tt.serviceExtensions)
 			require.Error(t, err)
 			assert.EqualError(t, err, tt.wantErrMsg)
 		})